% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urlshortener_objects.R
\name{AnalyticsSnapshot}
\alias{AnalyticsSnapshot}
\title{URL Shortener API Objects 
Lets you create, inspect, and manage goo.gl short URLs}
\usage{
AnalyticsSnapshot(browsers = NULL, countries = NULL, longUrlClicks = NULL,
  platforms = NULL, referrers = NULL, shortUrlClicks = NULL)
}
\arguments{
\item{browsers}{Top browsers, e}

\item{countries}{Top countries (expressed as country codes), e}

\item{longUrlClicks}{Number of clicks on all goo}

\item{platforms}{Top platforms or OSes, e}

\item{referrers}{Top referring hosts, e}

\item{shortUrlClicks}{Number of clicks on this short URL}
}
\value{
AnalyticsSnapshot object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2016-09-04 00:03:02
filename: /Users/mark/dev/R/autoGoogleAPI/googleurlshortenerv1.auto/R/urlshortener_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
AnalyticsSnapshot Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

