% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urlshortener_objects.R
\name{AnalyticsSummary}
\alias{AnalyticsSummary}
\title{AnalyticsSummary Object}
\usage{
AnalyticsSummary(allTime = NULL, day = NULL, month = NULL,
  twoHours = NULL, week = NULL)
}
\arguments{
\item{allTime}{Click analytics over all time}

\item{day}{Click analytics over the last day}

\item{month}{Click analytics over the last month}

\item{twoHours}{Click analytics over the last two hours}

\item{week}{Click analytics over the last week}
}
\value{
AnalyticsSummary object
}
\description{
AnalyticsSummary Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

