% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urlshortener_objects.R
\name{StringCount}
\alias{StringCount}
\title{StringCount Object}
\usage{
StringCount(count = NULL, id = NULL)
}
\arguments{
\item{count}{Number of clicks for this top entry, e}

\item{id}{Label assigned to this top entry, e}
}
\value{
StringCount object
}
\description{
StringCount Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

