% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urlshortener_functions.R
\name{url.get}
\alias{url.get}
\title{Expands a short URL or gets creation time and analytics.}
\usage{
url.get(shortUrl, projection = NULL)
}
\arguments{
\item{shortUrl}{The short URL, including the protocol}

\item{projection}{Additional information to return}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/urlshortener
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/urlshortener)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/url-shortener/v1/getting_started}{Google Documentation}
}

