% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urlshortener_functions.R
\name{url.insert}
\alias{url.insert}
\title{Creates a new short URL.}
\usage{
url.insert(Url)
}
\arguments{
\item{Url}{The \link{Url} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/urlshortener
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/urlshortener)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/url-shortener/v1/getting_started}{Google Documentation}

Other Url functions: \code{\link{Url}}
}
