% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urlshortener_objects.R
\name{Url}
\alias{Url}
\title{Url Object}
\usage{
Url(analytics = NULL, created = NULL, id = NULL, longUrl = NULL,
  status = NULL)
}
\arguments{
\item{analytics}{A summary of the click analytics for the short and long URL}

\item{created}{Time the short URL was created; ISO 8601 representation using the yyyy-MM-dd'T'HH:mm:ss}

\item{id}{Short URL, e}

\item{longUrl}{Long URL, e}

\item{status}{Status of the target URL}
}
\value{
Url object
}
\description{
Url Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Url functions: \code{\link{url.insert}}
}
