% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urlshortener_objects.R
\name{UrlHistory}
\alias{UrlHistory}
\title{UrlHistory Object}
\usage{
UrlHistory(items = NULL, itemsPerPage = NULL, nextPageToken = NULL,
  totalItems = NULL)
}
\arguments{
\item{items}{A list of URL resources}

\item{itemsPerPage}{Number of items returned with each full 'page' of results}

\item{nextPageToken}{A token to provide to get the next page of results}

\item{totalItems}{Total number of short URLs associated with this user (may be approximate)}
}
\value{
UrlHistory object
}
\description{
UrlHistory Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

