% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urlshortener_functions.R
\docType{package}
\name{urlshortener_googleAuthR}
\alias{urlshortener_googleAuthR}
\alias{urlshortener_googleAuthR-package}
\title{URL Shortener API
Lets you create, inspect, and manage goo.gl short URLs}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2016-09-04 00:03:02
filename: /Users/mark/dev/R/autoGoogleAPI/googleurlshortenerv1.auto/R/urlshortener_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/urlshortener
}
}

